Clean LocalFilesRepository class
